Fix trapezoidal motion profile for "inverted" profiles
Issue: Trapezoidal motion profile bug with initial and final velocities
at small distances. The "inverted" flag (i.e. whether to use a /\ or a
\/-shaped profile) was set only according to the difference in position
even when the correct result depended on initial and final velocity as
well.

Solution: Calculate the distance travelled by a linear velocity ramp
from the initial to the final velocity and compare it to the desired
distance. If it is smaller, invert the profile.