Add message viewer to webdash
I used to have standards, now I don't. Sorry :'(

This takes a lot of the logic for maintaining a list of queues from logging,
but exposes a function to get a message as JSON instead of as CSV.

The server sends over every most recent message as a single JSON object, which
is then parsed by the client-side javascript code.

This is all a horrific hack, and I'm really sorry about it. But it's better
than dealing with SmartDashboard...