Citrus robot in separate thread (#384)
* Move citrus robot to separate thread

I think test script fails but shhh

* Update joysticks in RobotPeriodic, fix thread priority (lol)

So it updates the joysticks in RobotPeriodic then reads from the joystick queues in citrus_robot/main.cpp. In theory. I also fixed the precious whitespace.

* Read driverstation info from a queue instead of DriverStation class

This supposedly makes it thread-safe

* Use a queue reader for driverstation queue

* Make rumble reusable and realtime

* Rename main.cpp/h to citrus_robot.cpp/h

Also adds that phased_loop thingy to not mess stuff up :)

* Make throttle_, wheel_, and joystick_ private

* Rename citrus_robot library to citrus_robot

* Change initializer for manipulator

* Remove unnecessary constructor

There should be no reason to create a joystick with no status queue but a rumble.

* Make Update function private, update comment

* Wesley's precious whitespace

* Move the joystick updating to someplace else