New Drivetrain Code (#648)
* Add drivetrain control library

This mega-commit includes:
 - A drivetrain model based on linear/angular kinematics and dynamics
 - A simple pose class, a pose class with curvature
 - A spline class
 - A trajectory class
 - A Non-linear Feedback Controller class for following a trajectory
 - Unit tests for all of the above

:P

* Add a generic drivetrain subsystem

This subsystem will be instatiated in all future robot codes. To adapt
to different drivetrains, a DrivetrainConfig struct will be customized.

This subsystem includes both an open-loop (blatant rip of CheesyDrive)
and a closed-loop (spline-following via a Nonlinear Feedback
controller).

There are no unit tests for a reason; unit testing a drivetrain without
(or even with) a model is too big a hassle to be worth it. Physical
tests will suffice.

* Lint, cleanup

* Fix signs in inverse dynamics

* Make the test more robust

The test now checks heading error as well as translational error. It
also simulates a helplessly underperforming robot (90% of whatever the
velocity goal actually is).